app/requirements.txt
fastapi==0.111.0
uvicorn==0.27.0
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
joblib>=1.3.0
nltk>=3.8.1
pytz
pycryptodome